+++ test/renderers/xml_renderer_test.rb
@@ -128,4 +128,10 @@ class TestXmlRenderer < ActiveSupport::TestCase
 
     assert_equal %q(<list><users type="array"/></list>), render_xml_output
   end
+
+  test "render underscorized xml" do
+    RablRails.xml_options = { :dasherize => false, :skip_types => false }
+    @template.source = { :first_name => :name }
+    assert_equal %q(<user><first_name>foobar</first_name></user>), render_xml_output
+  end
 end