+++ CHANGELOG.md
@@ -8,6 +8,7 @@
       RablRails.allow_empty_format_in_template = true
       RablRails.render(user, 'show') # => app/view/user.rabl
 
+  * Rails 4 support
   * Update travis configuration and remove warning in tests (petergoldstein)
 
 ## 0.3.3